kotlin/aoc2018/src/test/kotlin/Day04Tests.kt
import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.Test
import java.io.File

class Day04Tests {
    @Test
    fun `example part 1`() {
        // act
        val result = Day04().part1(listOf(
        "[1518-11-01 00:00] Guard #10 begins shift",
        "[1518-11-01 00:05] falls asleep",
        "[1518-11-01 00:25] wakes up",
        "[1518-11-01 00:30] falls asleep",
        "[1518-11-01 00:55] wakes up",
        "[1518-11-01 23:58] Guard #99 begins shift",
        "[1518-11-02 00:40] falls asleep",
        "[1518-11-02 00:50] wakes up",
        "[1518-11-03 00:05] Guard #10 begins shift",
        "[1518-11-03 00:24] falls asleep",
        "[1518-11-03 00:29] wakes up",
        "[1518-11-04 00:02] Guard #99 begins shift",
        "[1518-11-04 00:36] falls asleep",
        "[1518-11-04 00:46] wakes up",
        "[1518-11-05 00:03] Guard #99 begins shift",
        "[1518-11-05 00:45] falls asleep",
        "[1518-11-05 00:55] wakes up"
        ))

        // assert
        assertThat(result).isEqualTo(240)
    }

    @Test
    fun `do it part 1`() {
        // arrange
        val values = File(this.javaClass.getResource("input_day4").toURI()).readLines()


        // act
        val result = Day04().part1(values)

        // assert
        assertThat(result).isEqualTo(4716)
    }

    @Test
    fun `example part 2`() {
        // act
        val result = Day04().part2(listOf(
            "[1518-11-01 00:00] Guard #10 begins shift",
            "[1518-11-01 00:05] falls asleep",
            "[1518-11-01 00:25] wakes up",
            "[1518-11-01 00:30] falls asleep",
            "[1518-11-01 00:55] wakes up",
            "[1518-11-01 23:58] Guard #99 begins shift",
            "[1518-11-02 00:40] falls asleep",
            "[1518-11-02 00:50] wakes up",
            "[1518-11-03 00:05] Guard #10 begins shift",
            "[1518-11-03 00:24] falls asleep",
            "[1518-11-03 00:29] wakes up",
            "[1518-11-04 00:02] Guard #99 begins shift",
            "[1518-11-04 00:36] falls asleep",
            "[1518-11-04 00:46] wakes up",
            "[1518-11-05 00:03] Guard #99 begins shift",
            "[1518-11-05 00:45] falls asleep",
            "[1518-11-05 00:55] wakes up"
        ))

        // assert
        assertThat(result).isEqualTo(4455)
    }

    @Test
    fun `do it part 2`() {
        // arrange
        val values = File(this.javaClass.getResource("input_day4").toURI()).readLines()


        // act
        val result = Day04().part2(values)

        // assert
        assertThat(result).isEqualTo(117061)
    }
}